# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditEventRead": {
        "properties": {
          "action": {
            "description": "Ação auditada: CREATE, UPDATE, DELETE, CONNECT, DISCONNECT, etc.",
            "title": "Action",
            "type": "string"
          },
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "actor_name": {
            "title": "Actor Name",
            "type": "string"
          },
          "changes": {
            "additionalProperties": true,
            "description": "Dicionário sanitizado com valores antes e depois",
            "title": "Changes",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "id",
          "actor_id",
          "actor_name",
          "action",
          "entity_type",
          "entity_id",
          "created_at"
        ],
        "title": "AuditEventRead",
        "type": "object"
      },
...
      "GlobalSearchResponse": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/SearchGroup"
            },
            "title": "Groups",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "total_results": {
            "title": "Total Results",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "total_results",
          "groups"
        ],
        "title": "GlobalSearchResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImpactAnalysisRequest": {
        "properties": {
          "cable_segment_ids": {
            "description": "Lista de UUIDs dos trechos de cabo rompidos ou sob teste",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Cable Segment Ids",
            "type": "array"
          },
          "expected_topology_revision": {
            "description": "Revisão topológica esperada do cenário de rede",
            "title": "Expected Topology Revision",
            "type": "integer"
          }
        },
        "required": [
          "cable_segment_ids",
          "expected_topology_revision"
        ],
        "title": "ImpactAnalysisRequest",
        "type": "object"
      },
      "ImpactAnalysisResponse": {
        "properties": {
          "broken_segments_count": {
            "title": "Broken Segments Count",
            "type": "integer"
          },
          "impacted_ctos": {
            "description": "Lista de códigos de CTOs afetadas",
            "items": {
              "type": "string"
            },
            "title": "Impacted Ctos",
            "type": "array"
          },
          "impacted_customers": {
            "description": "Clientes ativos que perderam o enlace óptico",
            "items": {
              "$ref": "#/components/schemas/ImpactedCustomerItem"
            },
            "title": "Impacted Customers",
            "type": "array"
          },
          "impacted_pon_ports": {
            "description": "Portas PON afetadas pelo rompimento",
            "items": {
              "type": "string"
            },
            "title": "Impacted Pon Ports",
            "type": "array"
          },
          "previously_disconnected_count": {
            "description": "Clientes que já se encontravam desconectados antes do evento",
            "title": "Previously Disconnected Count",
            "type": "integer"
          },
          "topology_revision": {
            "title": "Topology Revision",
            "type": "integer"
          },
          "unaffected_customers_count": {
            "description": "Clientes em outros ramos que permaneceram com enlace ativo",
            "title": "Unaffected Customers Count",
            "type": "integer"
          },
          "unknown_status_count": {
            "description": "Clientes com documentação incompleta cujo impacto não pôde ser determinado",
            "title": "Unknown Status Count",
            "type": "integer"
          }
        },
        "required": [
          "topology_revision",
          "broken_segments_count",
          "impacted_customers",
          "unaffected_customers_count",
          "previously_disconnected_count",
          "unknown_status_count",
          "impacted_ctos",
          "impacted_pon_ports"
        ],
        "title": "ImpactAnalysisResponse",
        "type": "object"
      },
...
      "PaginatedResponse_AuditEventRead_": {
        "properties": {
          "items": {
            "description": "Lista de itens da página atual",
            "items": {
              "$ref": "#/components/schemas/AuditEventRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "Página retornada",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Tamanho de página aplicado",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total global de registros correspondentes",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedResponse[AuditEventRead]",
        "type": "object"
      },
...
      "PaginatedResponse_SplitterRead_": {
        "properties": {
          "items": {
            "description": "Lista de itens da página atual",
            "items": {
              "$ref": "#/components/schemas/SplitterRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "Página retornada",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Tamanho de página aplicado",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total global de registros correspondentes",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedResponse[SplitterRead]",
        "type": "object"
      },
...
      "SearchGroup": {
        "properties": {
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "entity_type",
          "items"
        ],
        "title": "SearchGroup",
        "type": "object"
      },
...
      "SplitterCreate": {
        "properties": {
          "code": {
            "description": "Código único do splitter (ex: SPL-CTO04-1x8)",
            "maxLength": 50,
            "minLength": 2,
            "title": "Code",
            "type": "string"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dispositivo onde o splitter está alojado (exclusivo com structure_id)",
            "title": "Device Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "output_ports_count": {
            "description": "Quantidade N de portas de saída (arquitetura estrita 1:N)",
            "maximum": 64.0,
            "minimum": 2.0,
            "title": "Output Ports Count",
            "type": "integer"
          },
          "ports": {
            "description": "Perdas reais medidas ou de datasheet cadastradas para cada saída em dB",
            "items": {
              "$ref": "#/components/schemas/SplitterPortLoss"
            },
            "title": "Ports",
            "type": "array"
          },
          "ratio": {
            "default": "1:8",
            "description": "Razão nominal de divisão (ex: 1:8, 1:16, 1:2 desbalanceado)",
            "title": "Ratio",
            "type": "string"
          },
          "structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Estrutura onde o splitter está instalado (exclusivo com device_id)",
            "title": "Structure Id"
          }
        },
        "required": [
          "code",
          "output_ports_count"
        ],
        "title": "SplitterCreate",
        "type": "object"
      },
...
      "SplitterRead": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_terminal_id": {
            "title": "Input Terminal Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "output_ports_count": {
            "title": "Output Ports Count",
            "type": "integer"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/SplitterPortRead"
            },
            "title": "Ports",
            "type": "array"
          },
          "ratio": {
            "title": "Ratio",
            "type": "string"
          },
          "structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "ratio",
          "output_ports_count",
          "input_terminal_id",
          "ports",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "SplitterRead",
        "type": "object"
      },
      "SplitterUpdate": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "ports": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SplitterPortLoss"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ports"
          }
        },
        "title": "SplitterUpdate",
        "type": "object"
      },
...
      "TraceDirection": {
        "enum": [
          "downstream",
          "upstream"
        ],
        "title": "TraceDirection",
        "type": "string"
      },
...
      "TraceRequest": {
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/TraceDirection",
            "default": "downstream",
            "description": "Direção da propagação do sinal óptico: downstream (PON->ONU) ou upstream (ONU->PON)"
          },
          "max_results": {
            "default": 50,
            "description": "Limite máximo de caminhos ou derivações retornados",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Results",
            "type": "integer"
          },
          "start_terminal_id": {
            "description": "UUID do terminal de início da travessia",
            "title": "Start Terminal Id",
            "type": "string"
          }
        },
        "required": [
          "start_terminal_id"
        ],
        "title": "TraceRequest",
        "type": "object"
      },
      "TraceResponse": {
        "properties": {
          "paths": {
            "description": "Caminhos ópticos válidos encontrados",
            "items": {
              "$ref": "#/components/schemas/TracePath"
            },
            "title": "Paths",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TraceStatus",
            "description": "Status de conclusão do rastreamento"
          },
          "topology_revision": {
            "description": "Revisão da topologia usada no rastreamento",
            "title": "Topology Revision",
            "type": "integer"
          },
          "unresolved_terminals": {
            "description": "Terminais com ponta aberta ou sem continuidade",
            "items": {
              "type": "string"
            },
            "title": "Unresolved Terminals",
            "type": "array"
          },
          "warnings": {
            "description": "Avisos e inconsistências detectadas",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "topology_revision",
          "status",
          "paths"
        ],
        "title": "TraceResponse",
        "type": "object"
      },
      "TraceStatus": {
        "enum": [
          "complete",
          "incomplete",
          "ambiguous",
          "cycle_detected",
          "limit_exceeded"
        ],
        "title": "TraceStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "FTTH Manager — Sistema de documentação física e óptica de rede FTTH",
    "title": "FTTH Manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/audit-events": {
      "get": {
        "description": "Retorna histórico ordenado de mutações e ações de usuários no sistema.",
        "operationId": "api_v1_audit_events_list_audit_events",
        "parameters": [
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AuditEventRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consultar trilha de auditoria append-only",
        "tags": [
          "Relatórios, Busca e Auditoria"
        ]
      }
    },
...
    "/api/v1/search": {
      "get": {
        "description": "Realiza busca textual indexada por código, nome ou serial com agrupamento por tipo de entidade.",
        "operationId": "api_v1_search_global_search",
        "parameters": [
          {
            "description": "Termo de pesquisa",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Termo de pesquisa",
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Limite máximo de resultados por grupo",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Limite máximo de resultados por grupo",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Busca global no inventário e rede",
        "tags": [
          "Relatórios, Busca e Auditoria"
        ]
      }
    },
...
    "/api/v1/splitters": {
      "get": {
        "operationId": "api_v1_splitters_list_splitters",
        "parameters": [
          {
            "description": "Filtrar por estrutura alojadora (CTO/CEO)",
            "in": "query",
            "name": "structure_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por estrutura alojadora (CTO/CEO)",
              "title": "Structure Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_SplitterRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar splitters",
        "tags": [
          "Splitters"
        ]
      },
      "post": {
        "operationId": "api_v1_splitters_create_splitter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitterRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cadastrar splitter",
        "tags": [
          "Splitters"
        ]
      }
    },
    "/api/v1/splitters/{splitter_id}": {
      "delete": {
        "operationId": "api_v1_splitters_splitter_id_delete_splitter",
        "parameters": [
          {
            "in": "path",
            "name": "splitter_id",
            "required": true,
            "schema": {
              "title": "Splitter Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Desativar splitter",
        "tags": [
          "Splitters"
        ]
      },
      "get": {
        "operationId": "api_v1_splitters_splitter_id_get_splitter",
        "parameters": [
          {
            "in": "path",
            "name": "splitter_id",
            "required": true,
            "schema": {
              "title": "Splitter Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitterRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalhes do splitter",
        "tags": [
          "Splitters"
        ]
      },
      "patch": {
        "operationId": "api_v1_splitters_splitter_id_update_splitter",
        "parameters": [
          {
            "in": "path",
            "name": "splitter_id",
            "required": true,
            "schema": {
              "title": "Splitter Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitterUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitterRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualizar splitter",
        "tags": [
          "Splitters"
        ]
      }
    },
...
    "/api/v1/topology/impact": {
      "post": {
        "description": "Simula a remoção virtual de trechos de cabos em um snapshot consistente da rede sem alterar dados operacionais, retornando os clientes e CTOs afetados.",
        "operationId": "api_v1_topology_impact_analyze_impact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpactAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImpactAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Análise de impacto de rompimento de cabos",
        "tags": [
          "Topologia e Grafo Óptico"
        ]
      }
    },
    "/api/v1/topology/trace": {
      "post": {
        "description": "Executa travessia consistente no grafo óptico a partir de um terminal de origem (PON ou ONU) com validação semântica de splitters e travessias internas.",
        "operationId": "api_v1_topology_trace_trace_path",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rastreamento óptico ponta a ponta",
        "tags": [
          "Topologia e Grafo Óptico"
        ]
      }
    },
...
  }
}